move depth tests outside of consolidate_integration
